refactor (#331)
